ai-agent/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "awsiotsdk>=1.28.1",
    "fastapi>=0.134.0",
    "langchain-anthropic>=1.3.4",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
